requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
pydantic>=2.6.0
python-multipart>=0.0.9
pandas>=2.2.0
python-jose[cryptography]
bcrypt
